move LateinitUsage rule to bugs package
