+++ README.md
@@ -416,8 +416,6 @@ style:
   active: true
   WildcardImport:
     active: true
-  LateinitUsage:
-    active: false
   NamingConventionViolation:
     active: true
     variablePattern: '^(_)?[a-z$][a-zA-Z$0-9]*$'
@@ -445,6 +443,8 @@ potential-bugs:
     active: true
   ExplicitGarbageCollectionCall:
     active: true
+  LateinitUsage:
+    active: false
 
 exceptions:
   active: true
