+++ detekt-cli/src/main/resources/default-detekt-config.yml
@@ -17,6 +17,8 @@ potential-bugs:
     active: true
   ExplicitGarbageCollectionCall:
     active: true
+  LateinitUsage:
+    active: false
 
 exceptions:
   active: true
@@ -99,8 +101,6 @@ style:
   active: true
   WildcardImport:
     active: true
-  LateinitUsage:
-    active: false
   MaxLineLength:
     active: true
     maxLineLength: 120
